refactor: Add CopyButton component for copying values
This commit adds a new component called CopyButton, which is responsible for copying values to the clipboard. It is used in the OTPGenerator component to provide a copy button for each value that needs to be copied. This improves the user experience by providing a convenient way to copy values without relying on the browser's default copy behavior.